# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/file:upload:
        put:
            tags:
                - OssService
            description: PUT方法上传文件
            operationId: OssService_PutUploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UploadOssFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadOssFileResponse'
        post:
            tags:
                - OssService
            description: POST方法上传文件
            operationId: OssService_PostUploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UploadOssFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UploadOssFileResponse'
...
    /admin/v1/files:
        get:
            tags:
                - FileService
            description: 查询文件列表
            operationId: FileService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: orderBy
                  in: query
                  description: 排序条件，其语法为JSON字符串，例如：{"val1", "-val2"}。字段名前加'-'为降序，否则为升序。
                  schema:
                    type: array
                    items:
                        type: string
                - name: query
                  in: query
                  description: AND过滤参数，其语法为json格式的字符串，如：{"key1":"val1","key2":"val2"}，具体请参见：https://github.com/tx7do/go-utils/tree/main/entgo/query/README.md
                  schema:
                    type: string
                - name: or
                  in: query
                  description: OR过滤参数，语法同AND过滤参数。
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListFileResponse'
        post:
            tags:
                - FileService
            description: 创建文件
            operationId: FileService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/files/{id}:
        get:
            tags:
                - FileService
            description: 查询文件详情
            operationId: FileService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/File'
        put:
            tags:
                - FileService
            description: 更新文件
            operationId: FileService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - FileService
            description: 删除文件
            operationId: FileService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateFileRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/File'
            description: 创建 - 请求
...
        File:
            type: object
            properties:
                id:
                    type: integer
                    description: 文件ID
                    format: uint32
                provider:
                    enum:
                        - MINIO
                        - ALIYUN
                        - AWS
                        - AZURE
                        - BAIDU
                        - QINIU
                        - TENCENT
                        - GOOGLE
                        - HUAWEI
                        - QCLOUD
                        - LOCAL
                    type: string
                    description: OSS供应商
                    format: enum
                bucketName:
                    type: string
                    description: 存储桶名称
                fileDirectory:
                    type: string
                    description: 文件目录
                fileGuid:
                    type: string
                    description: 文件Guid
                saveFileName:
                    type: string
                    description: 保存文件名
                fileName:
                    type: string
                    description: 文件名
                extension:
                    type: string
                    description: 文件扩展名
                size:
                    type: string
                    description: 文件字节长度
                sizeFormat:
                    type: string
                    description: 文件大小格式化
                linkUrl:
                    type: string
                    description: 链接地址
                md5:
                    type: string
                    description: md5码，防止上传重复文件
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 文件
...
        ListFileResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/File'
                total:
                    type: string
            description: 查询列表 - 回应
...
        UpdateFileRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/File'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新 - 请求
...
        UploadOssFileRequest:
            type: object
            properties:
                bucketName:
                    type: string
                    description: 文件桶名称
                objectName:
                    type: string
                    description: 文件名
                file:
                    type: string
                    description: 文件内容
                    format: bytes
                sourceFileName:
                    type: string
                    description: 原文件文件名
                mime:
                    type: string
                    description: 文件的MIME类型
        UploadOssFileResponse:
            type: object
            properties:
                url:
                    type: string
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...